Updates to adapter code as part of fluid banner testing
This commit updates the adapter code to work with fluid banner
testing. It sets a fluid banner ad size when the initial banner request
is received and converts that to a MREC size before the SDK load
method is called. When the adapter is about to call the SDK present
method, the adapter updates the presenting bannerView size to
the expected MREC size to prevent errors in the SDK. Comments
added to help with testing.